Draw pool into namespace. Draw commands nested in pool and made protected.
